Refactor: addCookie 메서드 통합
